+++ hosts/nixos/gammu/default.nix
@@ -9,6 +9,7 @@ in
     ./hardware-configuration.nix
     ../common-nixos.nix
     ../../../modules/common/certbot-secrets.nix
+    ../../../modules/common/development-packages.nix
     ../../../modules/nixos/syncthing-password.nix
     ../../../modules/nixos/syncthing-healthcheck.nix
   ];